+++ github-backend-dsgvo.php
@@ -24,11 +24,52 @@ function leafext_dsgvo_textdomain() {
 }
 add_action( 'plugins_loaded', 'leafext_dsgvo_textdomain' );
 
-if ( ! function_exists( 'leafext_get_repos' ) ) {
-	require_once LEAFEXT_DSGVO_PLUGIN_DIR . 'github/github-functions.php';
-}
-
-if ( is_main_site() && ! function_exists( 'leafext_update_puc_error' ) ) {
-	require_once LEAFEXT_DSGVO_PLUGIN_DIR . 'github/github-settings.php';
-	require_once LEAFEXT_DSGVO_PLUGIN_DIR . 'github/github-check-update.php';
+// Updates from Github
+if ( ! function_exists( 'leafext_updates_from_github' ) ) {
+	function leafext_updates_from_github() {
+		echo '<h2>' . wp_kses_post( 'Updates in WordPress way' ) . '</h2>';
+		if ( is_multisite() ) {
+			if ( strpos(
+				implode(
+					',',
+					array_keys(
+						get_site_option( 'active_sitewide_plugins', array() )
+					)
+				),
+				'leafext-update-github.php'
+			) !== false ) {
+						echo wp_kses_post(
+							'To manage and receive updates, open <a href="' .
+							get_site_url( get_main_site_id() ) .
+							'/wp-admin/admin.php?page=github-settings">Github settings</a>.'
+						);
+			} else {
+					echo wp_kses_post(
+						'To receive updates, go to the <a href="' .
+						esc_url( network_admin_url() ) .
+						'plugins.php">network dashboard</a> and install and network activate ' .
+						'<a href="https://github.com/hupe13/leafext-update-github">Updates for plugins from hupe13 hosted on Github</a>.'
+					);
+			}
+		} elseif ( strpos(
+			implode(
+				',',
+				get_option( 'active_plugins', array() )
+			),
+			'leafext-update-github.php'
+		) !== false ) {
+						echo wp_kses_post(
+							'To manage and receive updates, open <a href="' .
+							esc_url( admin_url() ) .
+							'admin.php?page=github-settings">Github settings</a>.'
+						);
+		} else {
+				echo wp_kses_post(
+					'To receive updates, go to the <a href="' .
+					esc_url( admin_url() ) .
+					'plugins.php">dashboard</a> and install and activate ' .
+					'<a href="https://github.com/hupe13/leafext-update-github">Updates for plugins from hupe13 hosted on Github</a>.'
+				);
+		}
+	}
 }